pull out epub renderers
this gives us a mechanism to create custom renderers for existing
pageblocks without having to actually modify those pageblocks
themselves.

Eg, we could create a renderer for QuizBlock that renders all quizzes as
if they have `rhetorical` set (making them in-place JS quizzes for the
purpose of the epub). That wouldn't make sense to have in QuizBlock in
general, but might be reasonable for a particular project.

Implementing that would be something along the lines of:

```
from pagetreeepub.renderers import BaseRenderer

class QuizBlockRenderer(BaseRenderer):
    def render(self, block):
		    block.rhetorical = True
				return block.render()

... then in settings ...

EPUB_BLOCK_RENDERERS = {
    'Quiz Block': myapp.renderers.QuizBlockRender,
}
```

(in practice, we'd probably want to do something more complicated,
pulling in different templates and such as well).